bgpd: sanity checks around issue with attr null
there are cases with ecmp where a crash happened when performing 'show
ip bgp' with ipv4 address family. This crash happens during qbgp
improvements, is not seen in quagga upstream. so this can be seen as a
regression for ipv4 address family. the fix consists in doing some
sanity checks that may no be all desirable. however this may help, and
until we find something better, we keep it.

Signed-off-by: Philippe Guibert <philippe.guibert@6wind.com>